+++ packages/react-native-the-sheet/src/keyboard-expander/index.tsx
@@ -1,2 +1,0 @@
-export * from './keyboard-expander'
-export * from './types'